Many fixes for ENG-707 to make recovery more robust in the face of multiple concurrent failures. Generally the approach taken is to fail the recovering node so that the rest of the cluster can successfully agree on failures. Also have several improvements for the fuzz test that make it better at generating load and validating data while failures and recovery are ongoing.

+++ src/frontend/org/voltdb/messaging/RecoveryMessage.java
@@ -124,7 +124,7 @@ public class RecoveryMessage extends VoltMessage {
     }
 
     public long txnId() {
-        return m_buffer.getInt(txnIdOffset);
+        return m_buffer.getLong(txnIdOffset);
     }
 
     public byte[] getMessageData() {
